refactor(launcher): extract window-lifecycle handlers into reducer/window.rs (task #182 PR-C)
Stacked on the re-applied pool extraction (which never reached main
due to PR #670 merging into the wrong base). Extracts 4 window-
lifecycle handlers from reducer/mod.rs into a new reducer/window.rs
sub-module.

Extracted:
- handle_report_window_opened (104 lines)
- handle_report_window_closed (51 lines)
- handle_report_backend_window_id_registered (20 lines)
- handle_report_backend_window_id_unregistered (20 lines)

Total ~195 lines moved out. mod.rs shrinks from 1007 → 815 lines.

## Cross-module wiring

- All 4 handlers marked pub(super) so only mod.rs's update() dispatcher
  can call them.
- host_is_running (helper in mod.rs) now pub(super) so window.rs can
  call it via super::host_is_running for OrphanInstance drift detection.
- Imports cleaned: WindowKind, HwndDriftKind, WindowMirror moved from
  mod.rs to window.rs (only place they're used now).
- tests.rs gets explicit imports for HwndDriftKind, WindowKind,
  WindowMirror (was relying on mod.rs's re-exports via super::*).
- One existing test that called handle_report_window_closed directly
  now calls super::window::handle_report_window_closed.

154 launcher tests pass.
