+++ agentmux-launcher/src/reducer/mod.rs
@@ -41,9 +41,9 @@
 //     register. Running → Quitting on Quit (B.3 placeholder).
 //     Quitting → Dead on cleanup-done (B.3 placeholder). No skipping.
 
-use agentmux_common::ipc::{ClientKind, Command, DriftKind, ErrorCode, Event, HwndDriftKind, WindowKind};
+use agentmux_common::ipc::{ClientKind, Command, DriftKind, ErrorCode, Event};
 
-use crate::state::{LifecyclePhase, ProcessRecord, ProcessState, State, WindowMirror};
+use crate::state::{LifecyclePhase, ProcessRecord, ProcessState, State};
 
 /// Context the reducer needs but can't read from State (clocks,
 /// connection identity). Passed in per-call so update remains pure.
@@ -74,6 +74,7 @@ pub struct Ctx {
 }
 
 mod pool;
+mod window;
 
 /// Apply one Command to State, returning the resulting Events. State
 /// is mutated in place. Total function — never panics on input
@@ -99,13 +100,13 @@ pub fn update(state: &mut State, cmd: Command, ctx: &Ctx) -> Vec<Event> {
             if let Some(err) = enforce_host_only(state, ctx, "ReportWindowOpened") {
                 return vec![err];
             }
-            handle_report_window_opened(state, ctx, label, kind, parent_label)
+            window::handle_report_window_opened(state, ctx, label, kind, parent_label)
         }
         Command::ReportWindowClosed { label } => {
             if let Some(err) = enforce_host_only(state, ctx, "ReportWindowClosed") {
                 return vec![err];
             }
-            handle_report_window_closed(state, label)
+            window::handle_report_window_closed(state, label)
         }
         Command::ReportPoolWindowAdded { label, saga_id } => {
             if let Some(err) = enforce_host_only(state, ctx, "ReportPoolWindowAdded") {
@@ -226,13 +227,13 @@ pub fn update(state: &mut State, cmd: Command, ctx: &Ctx) -> Vec<Event> {
             if let Some(err) = enforce_host_only(state, ctx, "ReportBackendWindowIdRegistered") {
                 return vec![err];
             }
-            handle_report_backend_window_id_registered(state, label, window_id)
+            window::handle_report_backend_window_id_registered(state, label, window_id)
         }
         Command::ReportBackendWindowIdUnregistered { label } => {
             if let Some(err) = enforce_host_only(state, ctx, "ReportBackendWindowIdUnregistered") {
                 return vec![err];
             }
-            handle_report_backend_window_id_unregistered(state, label)
+            window::handle_report_backend_window_id_unregistered(state, label)
         }
         // Phase B.9.1 (WRR) — Win32 reality events. Host-only
         // because only the host installs the SetWinEventHook /
@@ -513,45 +514,6 @@ pub fn update(state: &mut State, cmd: Command, ctx: &Ctx) -> Vec<Event> {
     }
 }
 
-/// Phase B.5 (window_id_map step a) — record the host-reported
-/// label → backend_window_id mapping. Idempotent on duplicate
-/// label (overwrites with the new ID and emits a fresh event so
-/// subscribers see the latest mapping).
-fn handle_report_backend_window_id_registered(
-    state: &mut State,
-    label: String,
-    window_id: String,
-) -> Vec<Event> {
-    state
-        .backend_window_ids
-        .insert(label.clone(), window_id.clone());
-    let v = state.bump_version();
-    vec![Event::BackendWindowIdRegistered {
-        label,
-        window_id,
-        version: v,
-    }]
-}
-
-/// Phase B.5 (window_id_map step a) — drop the host-reported label
-/// from the map. Strict pairing: emits `BackendWindowIdUnregistered`
-/// only when the label was present (mirrors `WindowClosed` and
-/// `PoolWindowRemoved` semantics — codex P2 PR #577 round-2).
-fn handle_report_backend_window_id_unregistered(
-    state: &mut State,
-    label: String,
-) -> Vec<Event> {
-    let removed = state.backend_window_ids.remove(&label);
-    let Some(window_id) = removed else {
-        return vec![];
-    };
-    let v = state.bump_version();
-    vec![Event::BackendWindowIdUnregistered {
-        label,
-        window_id,
-        version: v,
-    }]
-}
 
 
 /// Phase B.4 follow-up — drift check. Compares host-reported counts
@@ -671,165 +633,6 @@ fn enforce_host_only(state: &mut State, ctx: &Ctx, op: &'static str) -> Option<E
     })
 }
 
-/// Phase B.4 — record a host-reported window opening in the launcher's
-/// read-only mirror. Idempotent on duplicate opens (same label twice
-/// in a row): the second insert overwrites with fresh metadata and
-/// emits a fresh event. Subscribers must tolerate seeing the same
-/// label twice; cleaner once B.5 makes the launcher authoritative.
-///
-/// Phase B.5: also assigns an authoritative instance number from
-/// `state.instance_registry` and emits `WindowInstanceAssigned`.
-/// "main" is pre-seeded with 1; other labels get the next value of
-/// `next_instance_num`. Re-opens of an existing label preserve the
-/// original number — instance numbers are stable per-label-per-run.
-fn handle_report_window_opened(
-    state: &mut State,
-    ctx: &Ctx,
-    label: String,
-    kind: WindowKind,
-    parent_label: Option<String>,
-) -> Vec<Event> {
-    // PR #664 codex P1 round 2 — drain-on-WindowOpened RESTORED as
-    // best-effort fallback. The explicit `ReportHwndOpened(Some(label))`
-    // from `client.rs::on_after_created` remains the AUTHORITATIVE
-    // link, and `apply_hwnd_opened` REPAIRS stale links it finds.
-    // But that explicit dispatch is gated on `hwnd_val != 0` host-side;
-    // if the HWND can't be resolved at on_after_created time from
-    // either of the 2 sources (Views, host), the explicit dispatch
-    // is skipped and the mirror would otherwise stay permanently
-    // unlinked, breaking WRR drift detection AND orphan-destroy
-    // reconciliation (no WindowClosed when OS destroys the HWND →
-    // permanent ghost InstancePanel rows).
-    //
-    // (PR #664 round 4 dropped a 3rd fallback `find_own_top_level_window`
-    // because it returns the FIRST visible window in the process —
-    // some other window's HWND in a multi-window session — which
-    // would corrupt other labels' mirrors via the `Repaired` arm.
-    // See client.rs::on_after_created comment for details.)
-    //
-    // The drain provides a fallback link from `pending_hwnds`. If the
-    // drain picks the WRONG HWND (the original burst-create race), the
-    // subsequent `apply_hwnd_opened` call from on_after_created will
-    // detect the mismatch and REPAIR — see the `Repaired` arm there.
-    // Net: best-effort link via drain, authoritative repair via
-    // explicit dispatch. The combination addresses both the
-    // hwnd_val=0 case and the burst-create race.
-    const PENDING_AGE_LIMIT_MS: u64 = 2_000;
-    let drained_hwnd: Option<u64> = state
-        .pending_hwnds
-        .iter()
-        .filter(|(_, p)| p.label_hint.is_none())
-        .filter(|(_, p)| ctx.now_ms.saturating_sub(p.arrived_at_ms) <= PENDING_AGE_LIMIT_MS)
-        .max_by_key(|(_, p)| p.arrived_at_ms)
-        .map(|(hwnd, _)| *hwnd);
-    if let Some(hwnd) = drained_hwnd {
-        state.pending_hwnds.remove(&hwnd);
-    }
-
-    state.windows.insert(
-        label.clone(),
-        WindowMirror {
-            label: label.clone(),
-            kind,
-            parent_label: parent_label.clone(),
-            opened_at: ctx.now_rfc3339.clone(),
-            // Best-effort drain above; authoritative explicit
-            // ReportHwndOpened from on_after_created arrives a few
-            // ms later via `apply_hwnd_opened` and REPAIRS any wrong
-            // link the drain picked.
-            hwnd: drained_hwnd,
-            visible: false,
-            iconic: false,
-            last_rect: None,
-            last_foreground_at_ms: None,
-            foregrounded_since_open: false,
-        },
-    );
-    let mut out = Vec::with_capacity(2);
-    let v = state.bump_version();
-    out.push(Event::WindowOpened {
-        label: label.clone(),
-        kind,
-        parent_label,
-        version: v,
-    });
-
-    // Assign instance number if this label isn't already in the
-    // registry. Re-opens of an existing label keep the original
-    // number — matches host's `WindowInstanceRegistry` semantics
-    // where a label is only registered once per session.
-    let num = if let Some(existing) = state.instance_registry.get(&label).copied() {
-        existing
-    } else {
-        let n = state.next_instance_num;
-        state.instance_registry.insert(label.clone(), n);
-        state.next_instance_num += 1;
-        n
-    };
-    let v = state.bump_version();
-    out.push(Event::WindowInstanceAssigned { label, num, version: v });
-    out
-}
-
-/// Phase B.4 — drop a host-reported window from the mirror. Returns
-/// `Event::WindowClosed` only when the label was actually in the
-/// mirror; an unknown-label close is a silent no-op (codex P2 PR
-/// #577 round-2). Without this gate, a `ReportWindowClosed` for a
-/// label the launcher never saw (e.g. a pool window that was popped
-/// from the queue but failed HWND validation in
-/// `promote_pool_window` — the orphan window's eventual
-/// `on_before_close` reaches us without a matching open) would
-/// emit an unpaired `WindowClosed` broadcast and break subscribers
-/// that assume open/close pairing.
-///
-/// Phase B.5 — also drops the label from `instance_registry` and
-/// emits `WindowInstanceReleased` if a number was assigned.
-/// `next_instance_num` is NOT decremented — instance numbers are
-/// monotonic per-launcher-run.
-fn handle_report_window_closed(state: &mut State, label: String) -> Vec<Event> {
-    let was_present = state.windows.remove(&label).is_some();
-    if !was_present {
-        // Silent: only emit when the close pairs with a known open.
-        return vec![];
-    }
-    let mut out = Vec::with_capacity(4);
-    let v = state.bump_version();
-    out.push(Event::WindowClosed {
-        label: label.clone(),
-        version: v,
-        // Clean close — host ran on_before_close before sending
-        // ReportWindowClosed. F.6 saga is safe to trigger.
-        crash_detected: false,
-    });
-    if let Some(num) = state.instance_registry.remove(&label) {
-        let v = state.bump_version();
-        out.push(Event::WindowInstanceReleased { label: label.clone(), num, version: v });
-    }
-
-    // Phase B.9.3 — OrphanInstance transition. The label we just
-    // removed was the LAST user-visible window (state.windows is
-    // now empty). If a Host is still registered as Running, its
-    // own close path won't quit_message_loop because the warm
-    // pool is keeping state.browsers non-empty. Emit drift +
-    // saga-style HostShouldQuit so the host can reap pool and
-    // quit cleanly. See B.9.3 in
-    // docs/retro/next-steps-2026-04-29.md.
-    if state.windows.is_empty() && host_is_running(state) {
-        let v_drift = state.bump_version();
-        out.push(Event::HwndDriftDetected {
-            kind: HwndDriftKind::OrphanInstance,
-            label: Some(label),
-            hwnd: None,
-            detail: "Last user-visible window closed; host still alive (likely holding warm pool)"
-                .to_string(),
-            severity: agentmux_common::ipc::Severity::Warn,
-            version: v_drift,
-        });
-        let v_quit = state.bump_version();
-        out.push(Event::HostShouldQuit { version: v_quit });
-    }
-    out
-}
 
 /// Phase B.9.3 — does state.processes contain a Host in the
 /// `Running` lifecycle? Used by the OrphanInstance transition
@@ -837,7 +640,7 @@ fn handle_report_window_closed(state: &mut State, label: String) -> Vec<Event> {
 /// HostShouldQuit on every benign close. Tool-only sessions
 /// (`agentmux.exe --diag`) also correctly skip the saga because
 /// they never register a Host.
-fn host_is_running(state: &State) -> bool {
+pub(super) fn host_is_running(state: &State) -> bool {
     use crate::state::ProcessState;
     state.processes.values().any(|r| {
         r.kind == ClientKind::Host && matches!(r.state, ProcessState::Running)
